+++ cvat-core/src/server-proxy.ts
@@ -450,6 +450,13 @@ class ServerProxy {
         async function authorized() {
             try {
                 await getSelf();
+                // In CVAT app we use two types of authentication,
+                // So here we are forcing user have both credential types
+                // First request will fail if session is expired, then we check for precense of token
+                if (!store.get('token')) {
+                    await logout();
+                    return false;
+                }
             } catch (serverError) {
                 if (serverError.code === 401) {
                     removeToken();